Разбивает файл скрипта на модули
